Reorganizing files and fixing imports
