--- agr-normalizer/requirements.txt
@@ -1,11 +1,0 @@
-Flask==2.2.2
-Flask-Cors==3.0.10
-flask-restx==1.0.5
-SQLAlchemy==1.4.46
-alembic==1.9.2
-requests==2.28.2
-python-jose==3.3.0
-click==8.1.3
-xmlrunner==1.7.7
-environs==9.5.0
-psycopg2-binary==2.9.5
