+++ src/geneweaver/aon/app.py
@@ -2,7 +2,7 @@
 An entrypoint to application startup
 """
 
-from src.factory import create_app
+from geneweaver.aon.factory import create_app
 
 if __name__ == '__main__':
     app = create_app()
